Move `triangulate` to new `kernel::util` module
