+++ src/kernel/mod.rs
@@ -1,6 +1,7 @@
 pub mod geometry;
 pub mod shapes;
 pub mod topology;
+pub mod util;
 
 use parry3d_f64::bounding_volume::AABB;
