+++ src/kernel/topology/faces.rs
@@ -1,12 +1,10 @@
 use std::collections::BTreeSet;
 
 use decorum::R64;
-use nalgebra::point;
 use parry2d_f64::{
     bounding_volume::AABB,
     query::{Ray as Ray2, RayCast as _},
     shape::{Segment as Segment2, Triangle as Triangle2},
-    utils::point_in_triangle::{corner_direction, Orientation},
 };
 use parry3d_f64::{
     math::Isometry,
@@ -16,8 +14,7 @@ use parry3d_f64::{
 
 use crate::{
     debug::{DebugInfo, TriangleEdgeCheck},
-    kernel::geometry::Surface,
-    math::Point,
+    kernel::{geometry::Surface, util::triangulate},
 };
 
 use super::edges::Edges;
@@ -229,43 +226,4 @@ impl Face {
             Self::Triangles(triangles) => out.extend(triangles),
         }
     }
-}
-
-/// Create a Delaunay triangulation of all vertices
-pub fn triangulate(vertices: &[Point<2>]) -> Vec<Triangle2> {
-    use spade::Triangulation as _;
-
-    let points: Vec<_> = vertices
-        .iter()
-        .map(|vertex| spade::Point2 {
-            x: vertex.x,
-            y: vertex.y,
-        })
-        .collect();
-
-    let triangulation = spade::DelaunayTriangulation::<_>::bulk_load(points)
-        .expect("Inserted invalid values into triangulation");
-
-    let mut triangles = Vec::new();
-    for triangle in triangulation.inner_faces() {
-        let [v0, v1, v2] = triangle.vertices().map(|vertex| {
-            let pos = vertex.position();
-            point![pos.x, pos.y]
-        });
-
-        let triangle = match corner_direction(&v0, &v1, &v2) {
-            Orientation::Ccw => [v0, v1, v2].into(),
-            Orientation::Cw => [v0, v2, v1].into(),
-            Orientation::None => {
-                panic!(
-                    "Triangle returned from triangulation isn't actually a\
-                    triangle"
-                );
-            }
-        };
-
-        triangles.push(triangle);
-    }
-
-    triangles
 }
